refactored base classes to have as much of the boilerplate

+++ src/utils.py
@@ -21,13 +21,14 @@ def get_attention_mask(config: BenchmarkConfig) -> Dict[str, torch.Tensor]:
     attention_mask = torch.ones(
         config.batch_size,
         config.sequence_length,
-        dtype=torch.float,
+        dtype=torch.long,
         device=config.backend.device,
     )
-    
+
     if config.sparsity > 0:
         # apply sparse mask
-        mask = torch.rand((config.batch_size, config.sequence_length), device=config.backend.device)
+        mask = torch.rand(
+            (config.batch_size, config.sequence_length), device=config.backend.device)
         attention_mask[mask < config.sparsity] = 0
         attention_mask, _ = attention_mask.sort(dim=-1, descending=True)
